Move mshots functions from my-sites to lib
