+++ client/lib/mshots/index.js
